moved more xem routes from main to server
