+++ tests/test_server/test_server_routes_populate_queues.py
@@ -168,3 +168,47 @@ def test_send_single_activate_trigger_in_command__using_hex_notation__populates_
     assert response_json["ep_addr"] == 2
     assert response_json["bit"] == 1
     assert response_json["suppress_error"] is True
+
+
+def test_send_single_comm_delay_command__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+
+    expected_num_millis = 35
+    response = test_client.get(
+        f"/insert_xem_command_into_queue/comm_delay?num_milliseconds={expected_num_millis}"
+    )
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(0)
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "comm_delay"
+    assert communication["num_milliseconds"] == expected_num_millis
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "comm_delay"
+    assert response_json["num_milliseconds"] == expected_num_millis
+    assert response_json["suppress_error"] is True
+
+
+def test_send_single_get_num_words_fifo_command__populates_queue(
+    client_and_server_thread_and_shared_values,
+):
+    test_client, test_server_info, _ = client_and_server_thread_and_shared_values
+    test_server, _, _ = test_server_info
+    response = test_client.get("/insert_xem_command_into_queue/get_num_words_fifo")
+    assert response.status_code == 200
+
+    comm_queue = test_server.queue_container().get_communication_to_ok_comm_queue(0)
+    assert is_queue_eventually_not_empty(comm_queue) is True
+    communication = comm_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert communication["communication_type"] == "debug_console"
+    assert communication["command"] == "get_num_words_fifo"
+    assert communication["suppress_error"] is True
+    response_json = response.get_json()
+    assert response_json["command"] == "get_num_words_fifo"
+    assert response_json["suppress_error"] is True
